fsk-ans/flask_app/requirements.txt
flask
psutil
mysql-connector
